# src/Storybook.Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Storybook",
    "description": "",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "",
      "email": ""
    },
    "license": {
      "name": "Use under MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Campaign/{campaignId}": {
      "get": {
        "tags": [
          "Campaign"
        ],
        "summary": "Get Campaign by id.",
        "description": "Get Campaign by id.",
        "operationId": "getCampaignById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetCampaignByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Campaign"
        ],
        "summary": "Delete Campaign.",
        "description": "Delete Campaign.",
        "operationId": "removeCampaign",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveCampaignResponse"
            }
          }
        }
      }
    },
    "/api/Campaign": {
      "get": {
        "tags": [
          "Campaign"
        ],
        "summary": "Get Campaigns.",
        "description": "Get Campaigns.",
        "operationId": "getCampaigns",
        "produces": [
          "application/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetCampaignsResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Campaign"
        ],
        "summary": "Create Campaign.",
        "description": "Create Campaign.",
        "operationId": "createCampaign",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCampaignRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateCampaignResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Campaign"
        ],
        "summary": "Update Campaign.",
        "description": "Update Campaign.",
        "operationId": "updateCampaign",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateCampaignRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateCampaignResponse"
            }
          }
        }
      }
    },
    "/api/Campaign/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "Campaign"
        ],
        "summary": "Get Campaign Page.",
        "description": "Get Campaign Page.",
        "operationId": "getCampaignsPage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetCampaignsPageResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "CampaignDto": {
      "type": "object",
      "properties": {
        "campaignId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "CreateCampaignRequest": {
      "type": "object",
      "properties": {
        "campaign": {
          "$ref": "#/definitions/CampaignDto"
        }
      }
    },
    "CreateCampaignResponse": {
      "type": "object",
      "properties": {
        "campaign": {
          "$ref": "#/definitions/CampaignDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "GetCampaignByIdResponse": {
      "type": "object",
      "properties": {
        "campaign": {
          "$ref": "#/definitions/CampaignDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetCampaignsPageResponse": {
      "type": "object",
      "properties": {
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CampaignDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetCampaignsResponse": {
      "type": "object",
      "properties": {
        "campaigns": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CampaignDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
    "RemoveCampaignResponse": {
      "type": "object",
      "properties": {
        "campaign": {
          "$ref": "#/definitions/CampaignDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "UpdateCampaignRequest": {
      "type": "object",
      "properties": {
        "campaign": {
          "$ref": "#/definitions/CampaignDto"
        }
      }
    },
    "UpdateCampaignResponse": {
      "type": "object",
      "properties": {
        "campaign": {
          "$ref": "#/definitions/CampaignDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}